made FullTraversalQuery out of OdinsonEventQuery
